refactor: move system-memory probing from command layer to core/resources
