+++ src/core/resources/mod.rs
@@ -1,0 +1,3 @@
+//! System resource probing services.
+
+pub mod memory;
